Updates to re-use code for infrastrucutre set up
